ucm: Fix names to match with the driver
This changes the file and path names for the UCM rules to match up with
the information spit out by the ALSA driver. Effectively this makes the
UCM configs "exist" for pulseaudio.

This also removes and adjusts a few of the UCM rules to function
properly on the AIY board.

Change-Id: I2a887e296513844cd0afe24590688e54cc6656a9